[admin_api] move things from xds/internal to internal/xds
- package for env vars
- e2e test function for creating bootstrap file
